add clock demo

+++ src/includes/viewport.h
@@ -23,16 +23,14 @@ viewport_handle viewport_init(uint cols, uint rows, float aspect);
 
 void viewport_delete(viewport_handle handle);
 
-void viewport_draw_pixel(viewport_handle handle, uint x, uint y, float color);
+void viewport_draw_pixel(viewport_handle handle, int x, int y, float color);
 
 void viewport_render(viewport_handle handle);
 
-void viewport_draw_rectangle(viewport_handle handle,
-  uint x, uint y, uint w, uint h, float color);
+void viewport_draw_rectangle(viewport_handle handle, int x, int y, int w, int h, float color);
 
-void viewport_draw_line(viewport_handle handle, 
-  float px0, float py0, float px1, float py1, float color);
+void viewport_draw_line(viewport_handle handle, int px0, int py0, int px1, int py1, float color);
 
-void viewport_draw_ellipse(viewport_handle handle, float x, float y, float a, float b, float color);
+void viewport_draw_ellipse(viewport_handle handle, int x, int y, int a, int b, float color);
 
 #endif
